Accessor reworking progress
change capture pattern so all captured vars are scalar constants. not sure why i’m worried about capturing the capture object (if that’s even a problem), but it’s more explicit this way also.
change el look ups to always be passed an array (createdEls), which is always an array even though it will usually be a single element. needs more consideration.